Refactor GUI code to be in pypeit/gui. Move QA code out of core.
