--- pypeit/core/gui/__init__.py
@@ -1,5 +1,0 @@
-
-from pypeit.core.gui import identify
-from pypeit.core.gui import object_find
-from pypeit.core.gui import skysub_regions
-
