+++ pypeit/gui/__init__.py
@@ -1,0 +1,5 @@
+
+from pypeit.gui import identify
+from pypeit.gui import object_find
+from pypeit.gui import skysub_regions
+
