+++ pypeit/gui/object_find.py
