+++ pypeit/gui/setup_gui/images/window_icon.png
