+++ pypeit/scripts/identify.py
@@ -63,7 +63,7 @@ class Identify(scriptbase.ScriptBase):
         from pypeit import log
         from pypeit import PypeItError
         from pypeit.spectrographs.util import load_spectrograph
-        from pypeit.core.gui.identify import Identify
+        from pypeit.gui.identify import Identify
         from pypeit.wavecalib import BuildWaveCalib, WaveCalib
         from pypeit import slittrace
         from pypeit.images.buildimage import ArcImage
