+++ pypeit/wavecalib.py
@@ -16,11 +16,12 @@ from astropy.io import fits
 
 from pypeit import log
 from pypeit import PypeItError
-from pypeit.core import arc, qa
+from pypeit import qa
+from pypeit.core import arc
 from pypeit.core import fitting
 from pypeit.core import parse
 from pypeit.core.wavecal import autoid, wv_fitting, wvutils
-from pypeit.core.gui.identify import Identify
+from pypeit.gui.identify import Identify
 from pypeit import datamodel
 from pypeit import calibframe
 from pypeit.core.wavecal import echelle
