+++ pypeit/wavetilts.py
@@ -27,6 +27,7 @@ from pypeit.core import arc
 from pypeit.core import tracewave
 from pypeit.core.wavecal import autoid
 from pypeit.images import buildimage
+from pypeit.qa import arc_tilts_2d_qa, arc_tilts_spat_qa, arc_tilts_spec_qa
 
 
 class WaveTilts(calibframe.CalibFrame):
@@ -502,10 +503,33 @@ class BuildWaveTilts:
                 = tracewave.fit_tilts(trc_tilt_dict, thismask, slit_cen, spat_order=spat_order,
                                       spec_order=spec_order,maxdev=self.par['maxdev2d'],
                                       sigrej=self.par['sigrej2d'], func2d=self.par['func2d'],
-                                      doqa=doqa, calib_key=self.mstilt.calib_key,
-                                      slitord_id=self.slits.slitord_id[slit_idx],
-                                      minmax_extrap=self.par['minmax_extrap'],
-                                      show_QA=show_QA, out_dir=self.qa_path)
+                                      minmax_extrap=self.par['minmax_extrap'])
+
+        # Now do some QA
+        if doqa:
+            trc_tilt_dict_out = self.all_trace_dict[slit_idx]
+            calib_key=self.mstilt.calib_key
+            slitord_id=self.slits.slitord_id[slit_idx]
+            tilts_dspat = trc_tilt_dict_out['tilts_dspat']  # spatial offset from the central trace
+            tilts = trc_tilt_dict_out['tilts']  # legendre polynomial fit
+            tilts_2dfit = trc_tilt_dict_out['tilt_2dfit']
+            fwhm = trc_tilt_dict_out['fwhm']
+            tot_mask = trc_tilt_dict_out['tot_mask']
+            tilts_spec = trc_tilt_dict_out['tilts_spec']
+            rms_fit = trc_tilt_dict_out['rms_fit']
+            # Compute a rejection mask that we will use later. These are
+            # locations that were fit but were rejected
+            rej_mask = tot_mask & np.logical_not(trc_tilt_dict_out['fit_mask'])
+
+            # tot mask from the output mask
+            # rej mask and rms_fit not in any output, recompute, or put into output from fit_tilts?
+            arc_tilts_2d_qa(tilts_dspat, tilts, tilts_2dfit, tot_mask, rej_mask, spat_order, spec_order,
+                        rms_fit, fwhm, slitord_id=slitord_id, setup=calib_key, show_QA=show_QA, out_dir=self.qa_path)
+            arc_tilts_spat_qa(tilts_dspat, tilts, tilts_2dfit, tilts_spec, tot_mask, rej_mask, spat_order,
+                        spec_order, rms_fit, fwhm, slitord_id=slitord_id, setup=calib_key, show_QA=show_QA,
+                        out_dir=self.qa_path)
+            arc_tilts_spec_qa(tilts_spec, tilts, tilts_2dfit, tot_mask, rej_mask, rms_fit, fwhm,
+                        slitord_id=slitord_id, setup=calib_key, show_QA=show_QA, out_dir=self.qa_path)
 
         self.steps.append(inspect.stack()[0][3])
         return self.all_fit_dict[slit_idx]['coeff2']
